+++ outputs.tf
@@ -7,3 +7,13 @@ output "endpoint" {
   description = "endpoint for opensearch"
   value       = aws_opensearch_domain.this.endpoint
 }
+
+output "security_group_id" {
+  description = "Security group id for the opensearch."
+  value       = try(aws_security_group.this[0].id, "")
+}
+
+output "client_security_group_id" {
+  description = "Security group id for the opensearch client."
+  value       = try(aws_security_group.client[0].id, "")
+}