+++ variables.tf
@@ -61,7 +61,7 @@ variable "master_instance_type" {
   default     = "r6gd.large.search"
 
   validation {
-    condition     = can(regex("^[m3|r3|i3|i2|r6gd]", var.master_instance_type))
+    condition     = can(regex("^[t3|m3|r3|i3|i2|r6gd]", var.master_instance_type))
     error_message = "The EC2 master_instance_type must provide a SSD or NVMe-based local storage."
   }
 }
@@ -78,7 +78,7 @@ variable "hot_instance_type" {
   default     = "r6gd.large.search"
 
   validation {
-    condition     = can(regex("^[m3|r3|i3|i2|r6gd]", var.hot_instance_type))
+    condition     = can(regex("^[t3|m3|r3|i3|i2|r6gd]", var.hot_instance_type))
     error_message = "The EC2 hot_instance_type must provide a SSD or NVMe-based local storage."
   }
 }
@@ -182,6 +182,45 @@ variable "is_create_security_group" {
   description = "if true will create security group for opensearch"
   type        = bool
   default     = true
+}
+
+variable "additional_opensearch_security_group_ingress_rules" {
+  type = list(object({
+    from_port                = number
+    to_port                  = number
+    protocol                 = string
+    cidr_blocks              = list(string)
+    source_security_group_id = string
+    description              = string
+  }))
+  description = "Additional ingress rule for opensearch security group."
+  default     = []
+}
+
+variable "additional_opensearch_client_security_group_ingress_rules" {
+  type = list(object({
+    from_port                = number
+    to_port                  = number
+    protocol                 = string
+    cidr_blocks              = list(string)
+    source_security_group_id = string
+    description              = string
+  }))
+  description = "Additional ingress rule for opensearch client security group."
+  default     = []
+}
+
+variable "additional_opensearch_client_security_group_egress_rules" {
+  type = list(object({
+    from_port                = number
+    to_port                  = number
+    protocol                 = string
+    cidr_blocks              = list(string)
+    source_security_group_id = string
+    description              = string
+  }))
+  description = "Additional egress rule for opensearch client security group."
+  default     = []
 }
 
 variable "is_ebs_enabled" {